Feat: ensure product quantity can be decreased from cart

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -7,12 +7,12 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decraseCartProductQuantitySchema } from "./schema";
 
-export const removeProductFromCart = async (
-  data: z.infer<typeof removeProductFromCartSchema>,
+export const decraseCartProductQuantity = async (
+  data: z.infer<typeof decraseCartProductQuantitySchema>,
 ) => {
-  removeProductFromCartSchema.parse(data);
+  decraseCartProductQuantitySchema.parse(data);
   const session = await auth.api.getSession({
     headers: await headers(),
   });
@@ -25,12 +25,22 @@ export const removeProductFromCart = async (
       cart: true,
     },
   });
+
   if (!cartItem) {
     throw new Error("Product variant not found in cart");
   }
-  const cartDoesNotBelongToUser = cartItem.cart.userId != session.user.id
-  if(cartDoesNotBelongToUser){
+
+  const cartDoesNotBelongToUser = cartItem.cart.userId != session.user.id;
+  if (cartDoesNotBelongToUser) {
     throw new Error("Unauthorized");
   }
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+
+  if (cartItem.quantity == 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+  await db
+    .update(cartItemTable)
+    .set({ quantity: cartItem.quantity - 1 })
+    .where(eq(cartItemTable.id, cartItem.id));
 };
